+++ src/Ionization/IonizationFactory.h
@@ -16,7 +16,7 @@ public:
         std::string model=params.species_param[ispec].ionization_model;
 
         if ( model == "tunnel" ) {
-            if (params.species_param[ispec].charge > params.species_param[ispec].atomic_number)
+            if (params.species_param[ispec].charge > (int)params.species_param[ispec].atomic_number)
                 ERROR( "Charge > atomic_number for specie " << ispec );
 
             Ionize = new IonizationTunnel( params, ispec );
